+++ packages/plugin-testops/test/utils.ts
@@ -22,6 +22,7 @@ TestOpsClientMock.prototype = {
   stopUpload: vi.fn(),
   createLaunchCategoriesBulk: vi.fn().mockResolvedValue([]),
   closeLaunch: vi.fn(),
+  getNamedEnvFor: vi.fn().mockReturnValue(undefined),
 };
 
 export const AllureStoreMock = vi.fn(function () {});
